requirements.txt
# Django
Django
django-extensions

# Imagenes
Pillow

# ImDB API
imdbpy

# Debug
flake8